Clean architecture: splitting the monolith into layers
* Added the Application Core project. It will contain parts related to the
application domain: domain entities, user cases, and more.

* Added the Infrastructure project. This project will host the project
infrastructure. At the moment, it is planned to place there a
component responsible for authorization and authentication,
a service responsible for sending emails and a service responsible
for generating web tokens

* Added the Persistence project. This layer will be responsible
for accessing the application data

* Added a Web Api project. Represents the topmost layer of the
application. It will contain controllers for accessing the internal
layers of the application. Serves as an input / output tool for the
application