Web API: I / O layer
* Added a base controller class with a property for
accessing a mediator instance to run commands
 from the infrastructure and the application core.
The CQRS approach is used to separate the flow
of creating and retrieving data from the database.
The mediator pattern and the MediatR library are
used to loosen the connections between parts of
the application.

* Added api controller class with user registration action

* Added a middleware class for handling exceptions that
occur while the application is running.

* In the startup class, all the necessary parts of the
application are connected (the authorization and authentication
module, the data access module, services and the mediator
configuration), the request pipeline is configured, and the
exception handling middleware is connected