Feature: added registration functionality
* changed the user creation command. Removed the part
responsible for creating the user profile

* more specific exception classes are now generated when creating a user

* database connection strings are moved to a separate configuration file
appsettings. Infrastructure.json.

* added the interface and implementation of the user manager, which will
define functions that do not change the state of the user resource in the database.
These operations include sending a verification email, checking the validity of tokens,
checking the email and user name for uniqueness (for implementing asynchronous
validators on the client side) , etc.

* the endpoint responsible for registration has been changed. Now it combines the
operations of creating a user and sending an email to confirm the email address