Refactoring: changed exception handling middleware
* now, exceptions that occur in a particular component or layer of the
application will be handled in the appropriate method. Each component
or layer will have its own set of exceptions that are generated only in this
layer or component. All exceptions of the corresponding component or layer
will be inherited from a single base class, which will be used to form a
method that handles more specific exceptions. For example, all exceptions
of the Infrastructure component.Identities are inheritors of IdentityException.
Therefore, you can catch this exception and then define a specific exception
class in the HandleIdentityException() method of the ExceptionHandlerMiddleware
class and generate a response to the client

* changed the exception classes of Infrastructure.Identity. Now there is one base
class for all exception classes generated in this component. The inheritor classes
do not contain fields, since they do not extend the base class IdentityException, they
are needed to generate responses to the client for a specific error