Refactoring: changed the AppUser class
* it was still decided to divide the user into two entities: the identity entity
and the domain entity. Data such as first name, last name, city, country,
date of birth, etc. will be stored in the domain entity and, accordingly, in
another database. All identification data will be stored in a separate database.
The identity entity and the domain entity will be linked by a common ID