Feature: added email address verification functionality
* an action for confirming the email has been added to the AccontController. Why GET?
The method is marked with the HttpGet attribute, although in fact the resource (user) changes
(the value of the ConfirmedEmail flag changes). But the client is sent only a link to confirm the
email and I did not find an option in which it would be possible to specify what method the
request should have when clicking on the link. Perhaps there are some ways, but I could not
find them. I tried to mark the confirmEmail action of the Account controller with different attributes,
HttpPut and HttpPatch, but in this case the error code 405 was returned. Only the HttpGet option works

* added the ConfirmEmailCommand command. Why the command? When confirming an email
address, the value of the ConfirmedEmail flag for a specific user in the database changes,
so it was decided to display this action in a separate command