Feature: added password reset functionality
* password reset occurs in three steps
    - first, an email is sent to the user's email address with a link to reset the password.
The User Id and token are specified in the parameters of this Url. The endpoint of the request is
AccountController.SendResetPasswordEmail(), UserManagerService.SendResetPasswordEmail()
    - after clicking on the link, a request is sent to the server. The endpoint of the request is
 AccountController.VerifyResetPasswordToken(). In this action, the token is verified, then, upon
successful verification, this token and the user Id are recorded in cookies and returned to the
client, the client is redirected to a page with a form for entering a new password
    - after entering a new password, the user sends a request to the server with the new password
in the request body. The endpoint of the request is AccountController.ResetPassword(). In this action,
the user ID and token are extracted from the cookies, a new password is extracted from the request body,
and then this data is passed to the UserManagerService.ResetPasswordAsync()