Tests, refactoring:
* changed the abstract ApplicationDbContextFactory. Now, when creating a database,
a single AppUser test object with test data is entered into it. Added the Destroy method
that deletes the database after passing the tests

* added a base class for testing command handlers and queries. It has instance fields
of types  ApplicationDbContext and IMapper. Instances of the database context and mapper
are needed for tests. The Dispose () method is also defined, which calls the Destroy() method
of the abstract factory to delete the database after passing the tests.