Refactoring the password reset functionality
Password reset takes place in three stages:

      - sending an email to the user with a link to confirm the email address

      - after clicking on the link, a request is sent to the server to the end point of the token verification.
At this endpoint, the Id and password reset token are extracted from the Url. Then the token is
verified. If the verification was successful, the Id and token are recorded in the cookie and the client
is redirected to a page with a form for entering a new password. Otherwise, the client is redirected
to the ResetPasswordFailure page.

      - the user enters a new password and sends the form to the server. A new password is extracted from
the request body, and the Id and token are extracted from the cookies. Next, a new instance
of ResetPasswordCommand is created and its handler is executed. Depending on the outcome
of the reset the client is redirected to either the successful outcome page or the unsuccessful outcome page

Changes in the code:

* verification of the token from the password reset link:
      - the VerifyResetPasswordTokenAsync() method of the UserManagerService class now does not generate
an exception in case of an unsuccessful password reset. And the link for local redirection is formed depending
on the value returned by this method.
      - added the SetCookieValue(string key, string value) method to ApiControllerBase to add a value to cookies
      - added tests for the token verification endpoint

* reset password
      - added the ExtractCookieValue(string key) method to ApiControllerBase to extract the value from the cookie.
After extracting the value, the corresponding record is deleted from the cookies.
      - the ResetPasswordCommand handler now does not generate exceptions in case of an unsuccessful outcome,
but returns the corresponding boolean value, on the basis of which a link is formed for redirecting the client
      - added a test for the password reset endpoint